Change integers to decimals in migration, re-seed database with 3rd case.

+++ db/migrations/20190525185203_initial.js
@@ -3,14 +3,14 @@ exports.up = function(knex, Promise) {
   return Promise.all([
   	knex.schema.createTable('cases', function(table) {
   		table.increments('id').primary();
-  		table.integer('sodiumProductionRate');
+  		table.decimal('sodiumProductionRate');
   		table.decimal('potassiumProductionRate');
-  		table.integer('chlorideProductionRate');
-  		table.integer('bicarbonateProductionRate');
-  		table.integer('BUNProductionRate');
-  		table.integer('creatinineProductionRate');
-  		table.integer('calciumProductionRate');
-  		table.integer('filtrationFractionStarting');
+  		table.decimal('chlorideProductionRate');
+  		table.decimal('bicarbonateProductionRate');
+  		table.decimal('BUNProductionRate');
+  		table.decimal('creatinineProductionRate');
+  		table.decimal('calciumProductionRate');
+  		table.decimal('filtrationFractionStarting');
   		table.text('gender');
   		table.decimal('usualWeight');
   		table.text('historyOfPresentIllness');
